requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0

# AI/ML Dependencies
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
transformers==4.33.2
huggingface-hub==0.16.4

# PDF Processing
pdf2image==1.16.3
Pillow==10.0.0

# OCR
pytesseract==0.3.10

# Utilities
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.0.76
PyYAML==6.0.1

# Development (optional)
python-dotenv==1.0.0